+++ DotNet/Normalization/Domain/Entities/NormalizationConfig.cs
@@ -1,23 +1,16 @@
 ﻿#nullable disable
 using LantanaGroup.Link.Normalization.Domain.JsonObjects;
+using LantanaGroup.Link.Shared.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace LantanaGroup.Link.Normalization.Domain.Entities;
 
 [Table("NormalizationConfig")]
-public partial class NormalizationConfig
+public partial class NormalizationConfig : BaseEntityExtended
 {
-    [Key]
-    public int Id { get; set; }
-
     [Unicode(false)]
     public string FacilityId { get; set; }
 
     public Dictionary<string, INormalizationOperation> OperationSequence { get; set; }
-
-    public DateTime? CreatedDate { get; set; }
-
-    public DateTime? ModifiedDate { get; set; }
 }
